requirements.txt
mysql-connector-python
streamlit
SQLAlchemy
pandas
plotly


